+++ arcane/src/arcane/hdf5/srcs.cmake
@@ -5,8 +5,15 @@ set(ARCANE_SOURCES
   Hdf5VariableInfoBase.h
   Hdf5ReaderWriter.cc
   Hdf5ReaderWriter.h
+  Hdf5VariableReader.cc
+  Hdf5VariableWriter.cc
+  Hdf5MpiReaderWriter.cc
+  Hdf5MpiReaderWriter.h
 )
 
 set(AXL_FILES
   Hdf5ReaderWriter
+  Hdf5VariableReader
+  Hdf5VariableWriter
+  Hdf5MpiReaderWriter
 )
