+++ arcane/src/arcane/std/srcs.cmake
@@ -104,9 +104,6 @@ set(ARCANE_SOURCES
 )
 
 set(AXL_FILES
-  Hdf5VariableReader
-  Hdf5VariableWriter
-  Hdf5MpiReaderWriter
   EnsightHdfPostProcessor
   ArcaneBasicCheckpoint
   ArcaneLoadBalance
